Add a SpatialOptions component, refactor common options
