+++ src/components/scatterplot/ScatterplotOptions.js
@@ -1,53 +1,12 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Box from '@material-ui/core/Box';
 import Checkbox from '@material-ui/core/Checkbox';
 import Slider from '@material-ui/core/Slider';
-import Select from '@material-ui/core/Select';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
 import TableRow from '@material-ui/core/TableRow';
 import { capitalize } from '../../utils';
-
-const useStyles = makeStyles(theme => ({
-  box: {
-    boxSizing: 'border-box',
-  },
-  checkbox: {
-    padding: '3px',
-    color: theme.palette.primaryForeground,
-    '&:checked': {
-      color: theme.palette.primaryForeground,
-    },
-    '& input': {
-      height: '100%',
-    },
-  },
-  slider: {
-    color: theme.palette.primaryForeground,
-    minWidth: '60px',
-    padding: '10px 0 6px 0',
-  },
-  sliderValueLabel: {
-    '& span': {
-      '& span': {
-        color: theme.palette.primaryBackground,
-      },
-    },
-  },
-  tableContainer: {
-    overflow: 'hidden',
-  },
-  labelCell: {
-    padding: '2px 8px 2px 16px',
-  },
-  inputCell: {
-    padding: '2px 16px 2px 8px',
-    overflow: 'visible',
-  },
-}));
+import { useStyles } from '../shared-plot-options/styles';
+import OptionsContainer from '../shared-plot-options/OptionsContainer';
+import CellColorEncodingOption from '../shared-plot-options/CellColorEncodingOption';
 
 export default function ScatterplotOptions(props) {
   const {
@@ -84,99 +43,76 @@ export default function ScatterplotOptions(props) {
     setCellSetPolygonsVisible(event.target.checked);
   }
 
-  function handleColorEncodingChange(event) {
-    setCellColorEncoding(event.target.value);
-  }
-
   return (
-    <Box className={classes.box}>
-      <TableContainer className={classes.tableContainer}>
-        <Table className={classes.table} size="small">
-          <TableBody>
-            <TableRow>
-              <TableCell className={classes.labelCell}>
-                {observationsLabelNice} Set Labels Visible
-              </TableCell>
-              <TableCell className={classes.inputCell}>
-                <Checkbox
-                  className={classes.checkbox}
-                  checked={cellSetLabelsVisible}
-                  onChange={handleLabelVisibilityChange}
-                  name="scatterplot-option-cell-set-labels"
-                  color="default"
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell className={classes.labelCell}>
-                {observationsLabelNice} Set Label Size
-              </TableCell>
-              <TableCell className={classes.inputCell}>
-                <Slider
-                  disabled={!cellSetLabelsVisible}
-                  classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
-                  value={cellSetLabelSize}
-                  onChange={handleLabelSizeChange}
-                  aria-labelledby="cell-set-label-size-slider"
-                  valueLabelDisplay="auto"
-                  step={1}
-                  min={8}
-                  max={36}
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell className={classes.labelCell}>
-                {observationsLabelNice} Set Polygons Visible
-              </TableCell>
-              <TableCell className={classes.inputCell}>
-                <Checkbox
-                  className={classes.checkbox}
-                  checked={cellSetPolygonsVisible}
-                  onChange={handlePolygonVisibilityChange}
-                  name="scatterplot-option-cell-set-polygons"
-                  color="default"
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell className={classes.labelCell}>
-                {observationsLabelNice} Radius
-              </TableCell>
-              <TableCell className={classes.inputCell}>
-                <Slider
-                  classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
-                  value={cellRadius}
-                  onChange={handleRadiusChange}
-                  aria-labelledby="cell-radius-slider"
-                  valueLabelDisplay="auto"
-                  step={0.25}
-                  min={0.25}
-                  max={8}
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell className={classes.labelCell} htmlFor="cell-color-encoding-select">
-                {observationsLabelNice} Color Encoding
-              </TableCell>
-              <TableCell className={classes.inputCell}>
-                <Select
-                  native
-                  value={cellColorEncoding}
-                  onChange={handleColorEncodingChange}
-                  inputProps={{
-                    id: 'cell-color-encoding-select',
-                  }}
-                >
-                  <option value="cellSetSelection">Cell Sets</option>
-                  <option value="geneSelection">Gene Expression</option>
-                </Select>
-              </TableCell>
-            </TableRow>
-          </TableBody>
-        </Table>
-      </TableContainer>
-    </Box>
+    <OptionsContainer>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          {observationsLabelNice} Set Labels Visible
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Checkbox
+            className={classes.checkbox}
+            checked={cellSetLabelsVisible}
+            onChange={handleLabelVisibilityChange}
+            name="scatterplot-option-cell-set-labels"
+            color="default"
+          />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          {observationsLabelNice} Set Label Size
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Slider
+            disabled={!cellSetLabelsVisible}
+            classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
+            value={cellSetLabelSize}
+            onChange={handleLabelSizeChange}
+            aria-labelledby="cell-set-label-size-slider"
+            valueLabelDisplay="auto"
+            step={1}
+            min={8}
+            max={36}
+          />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          {observationsLabelNice} Set Polygons Visible
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Checkbox
+            className={classes.checkbox}
+            checked={cellSetPolygonsVisible}
+            onChange={handlePolygonVisibilityChange}
+            name="scatterplot-option-cell-set-polygons"
+            color="default"
+          />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          {observationsLabelNice} Radius
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Slider
+            classes={{ root: classes.slider, valueLabel: classes.sliderValueLabel }}
+            value={cellRadius}
+            onChange={handleRadiusChange}
+            aria-labelledby="cell-radius-slider"
+            valueLabelDisplay="auto"
+            step={0.25}
+            min={0.25}
+            max={8}
+          />
+        </TableCell>
+      </TableRow>
+      <CellColorEncodingOption
+        observationsLabel={observationsLabel}
+        cellColorEncoding={cellColorEncoding}
+        setCellColorEncoding={setCellColorEncoding}
+      />
+    </OptionsContainer>
   );
 }
