+++ src/components/spatial/SpatialSubscriber.js
@@ -11,6 +11,7 @@ import {
 } from '../data-hooks';
 import { getCellColors } from '../interpolate-colors';
 import Spatial from './Spatial';
+import SpatialOptions from './SpatialOptions';
 import SpatialTooltipSubscriber from './SpatialTooltipSubscriber';
 import { makeSpatialSubtitle, initializeLayerChannelsIfMissing, sortLayers } from './utils';
 import {
@@ -225,6 +226,13 @@ export default function SpatialSubscriber(props) {
       theme={theme}
       removeGridComponent={removeGridComponent}
       isReady={isReady}
+      options={(
+        <SpatialOptions
+          observationsLabel={observationsLabel}
+          cellColorEncoding={cellColorEncoding}
+          setCellColorEncoding={setCellColorEncoding}
+        />
+      )}
     >
       <Spatial
         ref={deckRef}
